Backend/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
pypdf==5.5.0
requests==2.32.3
python_dotenv==1.1.0
murf==2.0.2